+++ jcl/src/main/java/io/mypojo/jcl/utils/JclObjectFactory.java
@@ -16,8 +16,8 @@
 
 package io.mypojo.jcl.utils;
 
-import io.mypojo.jcl.Configuration;
 import io.mypojo.jcl.JarClassLoader;
+import io.mypojo.jcl.config.Configuration;
 import io.mypojo.jcl.exception.JclException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
