+++ jcl/src/test/resources/test/test.properties
@@ -13,4 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
 testkey=testval
